chore(banner): moves the banner function to the index
