+++ src/commands/index.ts
@@ -1,4 +1,3 @@
-import { CLICommand } from '../CLICommand';
 import '../parameter_declarations';
 import './create_drive';
 import './drive_info';
@@ -19,5 +18,3 @@ import './file_info';
 import './move_file';
 import './move_folder';
 import './get_drive_key';
-
-CLICommand.parse();
